% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binned_resid}
\alias{binned_resid}
\title{Binned Residual Plot}
\usage{
binned_resid(model, bins = 16)
}
\arguments{
\item{model}{the model object, which should have \code{fitted}, \code{resid}, etc. methods.}

\item{bins}{the number of bins}
}
\value{
A ggplot
}
\description{
Binned Residual Plot
}
