% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_helper.R
\name{calc_indiv_frame}
\alias{calc_indiv_frame}
\title{Create a model data frame for an individual respondent}
\usage{
calc_indiv_frame(row, nbhd, block_d, block_gr)
}
\arguments{
\item{row}{A single data frame row containing relevant individual covariates
for a respondent.}

\item{nbhd}{The respondent's neighborhood as a vector of indices indexing the
blocks in \code{block_gr}.}

\item{block_d}{A data frame of census blocks, including a column \code{centroid}
with an \code{s2} point geography of each block's centroid.}

\item{block_gr}{An adjacency graph object: a list containing an element
\code{graph} with the adjacency list, and \code{blocks} a character vector of GEOIDs
or codes corresponding to the indices in \code{graph}.}
}
\value{
A tibble that can be used inside a modeling function. Will contain
the entries in \code{row}, plus the relevant entries in \code{block_d} for each
block, plus columns:
\itemize{
\item \code{ring} containing the "ring" indicator around the residence: 0 indicates
the respondent's block, 1 for blocks touching the residential block, 2 for blocks
touching those, etc.
\item \code{incl} a binary indicator for whether the block is in the neighborhood
\item \code{dist} the distance to the respondent's block
\item \code{frac_con} the fraction of nearer blocks in the neighborhood this block
is connected to.
}
}
\description{
This function combines individual, neighborhood, and geographic information
to produce a data frame suitable for use in fitting a \code{\link[=neighborhood_model]{neighborhood_model()}}.
Can be applied in a loop over respondents to generate a model frame for
an entire sample.
}
