% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_calc.R
\name{eff_dist}
\alias{eff_dist}
\title{Get Posterior Mean of Effective Block Distance}
\usage{
eff_dist(fit, new_resp, block_d, proc_fn = function(x) x)
}
\arguments{
\item{fit}{the model fit, from \code{\link[=neighborhood_model]{neighborhood_model()}}}

\item{new_resp}{a single-row respondent data frame to make predictions from}

\item{block_d}{the block data frame}

\item{proc_fn}{a processing function that is used to prepare raw model data
for fitting}
}
\value{
a numeric vector of effective distances
}
\description{
Not exported. Assumes \code{block_d} has column \code{fips} which is used inside
the \code{neighborhood} column of \code{new_resp}..
}
