% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_area}
\alias{local_area}
\title{Filter block data to a radius of a FIPS code}
\usage{
local_area(fips, block_d, geom_d, dist = 0.5)
}
\arguments{
\item{fips}{the FIPS code to center the area at}

\item{block_d}{the block data, with a \verb{$fips} column matching \code{fips} argument.}

\item{geom_d}{the block geometry data}

\item{distance}{the radius of the area, in miles}
}
\description{
Filter block data to a radius of a FIPS code
}
