% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{nbhd_fit}
\alias{nbhd_fit}
\alias{summary.nbhd_fit}
\alias{coef.nbhd_fit}
\alias{fixef.nbhd_fit}
\alias{ranef.nbhd_fit}
\alias{fitted.nbhd_fit}
\alias{residuals.nbhd_fit}
\alias{as.matrix.nbhd_fit}
\alias{as.data.frame.nbhd_fit}
\title{Functions for working with neighborhood fits}
\usage{
\method{summary}{nbhd_fit}(object, ...)

\method{coef}{nbhd_fit}(object, ...)

\method{fixef}{nbhd_fit}(object, ...)

\method{ranef}{nbhd_fit}(object, ...)

\method{fitted}{nbhd_fit}(object, ...)

\method{residuals}{nbhd_fit}(object, ...)

\method{as.matrix}{nbhd_fit}(x, ...)

\method{as.data.frame}{nbhd_fit}(x, ...)
}
\arguments{
\item{object, x}{a \code{nbhd_fit} object}

\item{...}{Ignored.}
}
\description{
Functions for working with neighborhood fits
}
