% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbhdmodel-package.R
\docType{package}
\name{nbhdmodel-package}
\alias{nbhdmodel}
\alias{nbhdmodel-package}
\title{nbhdmodel: Neighborhood Modeling and Analysis}
\description{
Functionality for fitting neighborhood models of McCartan, Brown, and Imai \href{https://arxiv.org/abs/2110.14014}{arXiv:2110.14014}. The core methodology is described in the paper and can be implemented with any tool that can fit generalized linear mixed models (GLMMs). However, some of the preprocessing necessary to set up the GLMM can be onerous. In addition to providing a specialized GLMM routine, this package provides several preprocessing functions that, while not completely general, should be useful for others performing these kinds of analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://corymccartan.github.io/nbhdmodel}
  \item \url{https://github.com/CoryMcCartan/nbhdmodel}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu}

Authors:
\itemize{
  \item Jacob Brown \email{jrbrown@g.harvard.edu}
  \item Kosuke Imai \email{imai@harvard.edu}
}

}
\keyword{internal}
