% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{neighborhood_model}
\alias{neighborhood_model}
\title{Fit the Neighborhood Model}
\usage{
neighborhood_model(
  formula,
  data,
  prior_coef_scale = 2.5,
  draws = 1000,
  imp_samp = TRUE,
  init = 0,
  ...,
  hessian = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{formula}{a one-sided model formula. The actual formula used to fit the
GLMM will be generated from this one: it will contain a term for log(dist)
(where \code{dist} must be the column that encodes distance), and will have random
effects based on the \code{id} column.  The column \code{incl} will be used as the
left-hand-side variable indicating that a block is in the neighborhood.}

\item{data}{the model data frame; usually the output of \code{\link[=make_model_frame]{make_model_frame()}}.
Should have an \code{incl} column for block inclusion, an  \code{id} column with
respondent IDs, and a \code{dist} column with distances.}

\item{prior_coef_scale}{the scale of the prior on the standardized predictors.}

\item{draws}{the number of approximate posterior draws to generate}

\item{imp_samp}{whether to perform importance resampling on the approximate draws.}

\item{init}{initial values for the model fitting function \code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}}.}

\item{...}{other arguments to to the model fitting function \code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}}.}

\item{hessian}{whether to compute the Hessian. Required for full inference.}

\item{verbose}{if \code{TRUE}, show verbose optimization output.}
}
\value{
a fitted model object of class \link{nbhd_fit}, which is a list which
includes the following elements:
\itemize{
\item \code{map}, the MAP estimates for the parameters.
\item \code{vcov}, the covariance matrix for the MAP estimates, calculated from the
Hessian of the log posterior.
\item \code{raw_ids} the vector of ids
\item \code{X} the design matrix
\item \code{y} the outcome value (1 - \code{incl})
\item \code{post}, the approximate posterior samples, from \link[posterior:draws_rvars]{posterior::draws_rvars}
\item \code{lp}, the log posterior probability of each sample
\item \code{lp_norm}, the log probability of Normal approx. to posterior for each sample
}
}
\description{
Fits the neighborhood GLMM using a provided formula and using a
Bernoulli outcome with cloglog link.
}
