% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{plot.nbhd_fit}
\alias{plot.nbhd_fit}
\title{Plot coefficient estimates}
\usage{
\method{plot}{nbhd_fit}(x, y = NULL, intercept = FALSE, inner_prob = 0.5, outer_prob = 0.9, ...)
}
\arguments{
\item{x}{a \code{nbhd_fit} object from \code{\link[=neighborhood_model]{neighborhood_model()}}.}

\item{y}{ignored}

\item{intercept}{if \code{FALSE}, don't plot the intercept estimate.}

\item{inner_prob}{the inner credible interval probability}

\item{outer_prob}{the inner credible interval probability}

\item{...}{Ignored.}
}
\value{
A ggplot.
}
\description{
50\% and 90\% credible intervals plotted by default.
}
