% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_calc.R
\name{post_incl}
\alias{post_incl}
\title{Get Posterior Block Inclusion Probabilities}
\usage{
post_incl(
  fit,
  new_resp,
  resp_id = NULL,
  block_d,
  proc_fn = function(x) x,
  use_distance = TRUE
)
}
\arguments{
\item{fit}{the model fit, from \code{\link[=neighborhood_model]{neighborhood_model()}}}

\item{new_resp}{a single-row respondent data frame to make predictions from}

\item{resp_id}{the resondent ID, used to get the random effects. If \code{NULL}
a new random effect will be simulated for each draw. If \code{NA}, the random
effect will be set to zero.}

\item{block_d}{the block data frame}

\item{proc_fn}{a processing function that is used to prepare raw model data
(from \code{\link[=make_model_frame]{make_model_frame()}}) for fitting}

\item{use_distance}{if \code{FALSE}, remove block-to-home distance from the linear
predictor.}
}
\value{
a numeric vector of inclusion probabilities
}
\description{
Get Posterior Block Inclusion Probabilities
}
