% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nbhd.R
\name{simulate_neighborhood}
\alias{simulate_neighborhood}
\title{Simulate a neighborhood for a respondent}
\usage{
simulate_neighborhood(
  fit,
  new_resp,
  draws = 1,
  resp_id = NULL,
  block_d,
  block_gr,
  max_ring = 10L,
  proc_fn = function(x) x
)
}
\arguments{
\item{fit}{the model fit, from \code{\link[=neighborhood_model]{neighborhood_model()}}}

\item{new_resp}{a single-row respondent data frame to make predictions from.
Should have a \code{neighborhood} column with codes matching \code{block_gr$blocks}.}

\item{draws}{the number of simulated neighborhoods per respondent}

\item{resp_id}{the resondent ID, used to get the random effects. If \code{NULL}
a new random effect will be simulated for each draw. If \code{NA}, the random
effect will be set to zero. If a numeric, the random effect will be set to
this value.}

\item{block_d}{the block data frame. Should have a \code{centroid} geography
column.}

\item{block_gr}{An adjacency graph object: a list containing an element
\code{graph} with the adjacency list, and \code{blocks} a character vector of GEOIDs
or codes corresponding to the indices in \code{graph}.}

\item{max_ring}{the maximum graph distance from the starting block to allow.}

\item{proc_fn}{a processing function that is used to prepare raw model data
(from \code{\link[=make_model_frame]{make_model_frame()}}) for fitting}
}
\value{
A list of \code{draws} integer vectors containing the indices of the
blocks (in \code{block_d}) making up each simulated neighborhood.
}
\description{
Simulate a neighborhood for a respondent
}
